added scene changes
