+++ SPOOP/Assets/Scripts/Character Creation.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: df9006df1653aa0499da99a61ffc1f6c
+folderAsset: yes
+timeCreated: 1519162379
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
