+++ SPOOP/Assets/Scripts/Character Creation/CharacterCreationNextLevel.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 9ecd23e1a1c683043b183276880e7b07
+timeCreated: 1519162417
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
