+++ SPOOP/Assets/Scripts/Main Menu.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 61115da596cdde345979c81d7591c09c
+folderAsset: yes
+timeCreated: 1519161954
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
