+++ SPOOP/Assets/Scripts/Main Menu/MainMenuNextLevel.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class MainMenuNextLevel : MonoBehaviour
+{
+ 
+    public void OnMouseButton()
+    {
+        SceneManager.LoadScene("Character Creation");
+    }
+}
