+++ SPOOP/Assets/Scripts/Main Menu/MainMenuQuitLevel.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MainMenuQuitLevel : MonoBehaviour
+{
+
+    public void OnMouseButton()
+
+    {
+        Application.Quit();
+    }
+}
+
+
