+++ SPOOP/Assets/Scripts/Main Menu/MainMenuQuitLevel.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: bc9793442aa29654e8242c05cc31cfdc
+timeCreated: 1519161390
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
